feat: Add common SHA3 algorithms to the random data generator editor action
